Update zram-install.sh
Almost the entire script has been revised to be more transparent and let the user decide which compression algorithm to use and how much RAM to use. Hints have also been added as to which algorithm seems to be the best and which the worst choice, as well as a recommendation as to how much RAM should be used. In addition, all algorithms currently available or implemented in current kernels have been added.