Update topcmd
The function now determines the operating system (func topcmd_os) and take into account, that MacOS and FreeBSD used to have a command "column", that doesn't know headers. As a result the output under Linux looks a little bit nicer.

Function _pager was added. With this function it is possible to distinguish between and configure different pagers.

Furthermore under Linux the command "fc" will be used instead of the history command, because the latter is just an alias of "fc -l". Now it is more transparent.

The help section is now much more comprehensive. Some will say it is too much of a good thing.